+++ src/scandeval/benchmark_config_factory.py
@@ -132,9 +132,17 @@ def build_benchmark_config(
     if use_flash_attention is None:
         use_flash_attention = importlib.util.find_spec("flash_attn") is not None
         if not use_flash_attention:
-            logger.debug(
-                "Flash Attention has not been installed, so this will not be used."
+            message = (
+                "Flash Attention has not been installed, so this will not be used. "
+                "To install it, run `pip install -U wheel && "
+                "FLASH_ATTENTION_SKIP_CUDA_BUILD=TRUE pip install flash-attn "
+                "--no-build-isolation`. You can disable this message by setting "
             )
+            if run_with_cli:
+                message += "`--use-flash-attention` or `--no-use-flash-attention`."
+            else:
+                message += "`use_flash_attention=True/False` in the `Benchmarker`."
+            logger.info(message)
 
     return BenchmarkConfig(
         model_languages=model_languages,